Moved influxdb-pojomapper exception to parent package #340
